www/requirements.txt
Cheetah==2.4.4
decorator==4.0.10
Django==1.9.7
django-cors-headers==1.1.0
django-suit==0.2.19
elasticsearch==2.3.0
elasticsearch-dsl==2.1.0
jsmin==2.2.1
kafka-python==1.2.4
kazoo==2.2.1
List==1.3.0
lxml==3.5.0
Markdown==2.6.6
markdown2==2.3.1
networkx==1.11
psutil==4.3.0
pyftpdlib==1.5.1
python-dateutil==2.5.3
PyYAML==3.11
six==1.10.0
urllib3==1.16
zc.thread==1.0.0
zc.zk==2.1.0
